auto weights download fix for Windows

+++ utils/google_utils.py
@@ -4,6 +4,7 @@
 
 import os
 import time
+import requests
 from pathlib import Path
 
 
@@ -40,35 +41,70 @@ def gdrive_download(id='1n_oKgR81BJtqk75b00eAjdv03qVCQn2f', name='coco128.zip'):
     # Downloads a file from Google Drive, accepting presented query
     # from utils.google_utils import *; gdrive_download()
     t = time.time()
+    if os.name == 'nt':
+        r = windows_download(id, name)
+    else:
+        print('Downloading https://drive.google.com/uc?export=download&id=%s as %s... ' % (id, name), end='')
+        os.remove(name) if os.path.exists(name) else None  # remove existing
+        os.remove('cookie') if os.path.exists('cookie') else None
 
-    print('Downloading https://drive.google.com/uc?export=download&id=%s as %s... ' % (id, name), end='')
-    os.remove(name) if os.path.exists(name) else None  # remove existing
-    os.remove('cookie') if os.path.exists('cookie') else None
+        # Attempt file download
+        os.system("curl -c ./cookie -s -L \"drive.google.com/uc?export=download&id=%s\" > /dev/null" % id)
+        if os.path.exists('cookie'):  # large file
+            s = "curl -Lb ./cookie \"drive.google.com/uc?export=download&confirm=`awk '/download/ {print $NF}' ./cookie`&id=%s\" -o %s" % (
+                id, name)
+        else:  # small file
+            s = 'curl -s -L -o %s "drive.google.com/uc?export=download&id=%s"' % (name, id)
+        r = os.system(s)  # execute, capture return values
+        os.remove('cookie') if os.path.exists('cookie') else None
 
-    # Attempt file download
-    os.system("curl -c ./cookie -s -L \"drive.google.com/uc?export=download&id=%s\" > /dev/null" % id)
-    if os.path.exists('cookie'):  # large file
-        s = "curl -Lb ./cookie \"drive.google.com/uc?export=download&confirm=`awk '/download/ {print $NF}' ./cookie`&id=%s\" -o %s" % (
-            id, name)
-    else:  # small file
-        s = 'curl -s -L -o %s "drive.google.com/uc?export=download&id=%s"' % (name, id)
-    r = os.system(s)  # execute, capture return values
-    os.remove('cookie') if os.path.exists('cookie') else None
+        # Error check
+        if r != 0:
+            os.remove(name) if os.path.exists(name) else None  # remove partial
+            print('Download error ')  # raise Exception('Download error')
+            return r
 
-    # Error check
-    if r != 0:
-        os.remove(name) if os.path.exists(name) else None  # remove partial
-        print('Download error ')  # raise Exception('Download error')
-        return r
-
-    # Unzip if archive
-    if name.endswith('.zip'):
-        print('unzipping... ', end='')
-        os.system('unzip -q %s' % name)  # unzip
-        os.remove(name)  # remove zip to free space
+        # Unzip if archive
+        if name.endswith('.zip'):
+            print('unzipping... ', end='')
+            os.system('unzip -q %s' % name)  # unzip
+            os.remove(name)  # remove zip to free space
 
     print('Done (%.1fs)' % (time.time() - t))
     return r
+
+
+def windows_download(id, name):
+    try:
+        URL = "https://docs.google.com/uc?export=download"
+        CHUNK_SIZE = 32768
+
+        session = requests.Session()
+
+        response = session.get(URL, params = { "id" : id }, stream = True)
+        token = get_confirm_token(response)
+
+        if token: # required for big files
+            params = { "id" : id, "confirm" : token }
+            response = session.get(URL, params = params, stream = True)
+
+        with open(name, "wb") as f:
+            for chunk in response.iter_content(CHUNK_SIZE):
+                if chunk:
+                    f.write(chunk)
+
+        return 0 # all good
+
+    except Exception as e:
+        print("Got exception %s when trying to download %s" % (e, name))
+        return 'not 0' # somehting went wrong
+
+def get_confirm_token(response):
+    for key, value in response.cookies.items():
+        if key.startswith('download_warning'):
+            return value
+
+    return None
 
 
 # def upload_blob(bucket_name, source_file_name, destination_blob_name):
